requirements.txt
appdirs==1.4.3
benford-py==0.3.3
Brotli==1.0.9
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
dash==1.19.0
dash-bootstrap-components==0.11.1
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
detect-delimiter==0.1.1
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
lockfile==0.12.2
MarkupSafe==1.1.1
matplotlib==3.3.3
msgpack==0.6.2
numpy==1.19.5
packaging==20.3
pandas==1.2.1
pep517==0.8.2
Pillow==8.1.0
plotly==4.14.3
progress==1.5
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.5
requests==2.22.0
retrying==1.3.3
six==1.14.0
SQLAlchemy==1.3.22
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
